qiling_win/requirements.txt
qiling==1.2.3
yara-python==4.0.5
karton-core==4.2.0
